refactor(release): move the start-position verdict into _lib
Two CI gates, one cause: `release.ts` sits 456 lines past the 1500-line source
ratchet, where every added line is one unit of excess, so the verdict plus its
rationale pushed `check_source_size_budget` 85 over its baseline — and raising
the baseline is a defect, not a fix. `_lib/release_position.ts` is 96 lines and
therefore under the cap, so the same prose costs nothing; `release.ts` ends at
1869, two lines below where it started.

`lint_code_comments` separately refused the `##` markdown headings in the
docblock as report structure in source. Rewritten as plain prose; the content
is unchanged.

The wiring test now also asserts the import, so the extraction cannot silently
become a second dead copy.
